% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_add_package.R
\name{pkgr_add_package}
\alias{pkgr_add_package}
\title{Add Package(s) to Repository}
\usage{
pkgr_add_package(package_files, repo_path)
}
\arguments{
\item{package_files}{Character vector of package files.}

\item{repo_path}{Repository path.}
}
\value{
Repository path.
}
\description{
Copies package files into a repository and
automatically updates repository metadata.
}
