% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_build_bundle.R
\name{pkgr_build_bundle}
\alias{pkgr_build_bundle}
\title{Build Deployment Bundle}
\usage{
pkgr_build_bundle(package, repo_path, bundle_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Package repository path.}

\item{bundle_path}{Bundle output path.}
}
\value{
Bundle path.
}
\description{
Creates a deployment bundle containing
required package files and deployment manifest.
}
