% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_build_library.R
\name{pkgr_build_library}
\alias{pkgr_build_library}
\title{Build Library}
\usage{
pkgr_build_library(package_file, library_path)
}
\arguments{
\item{package_file}{Package file.}

\item{library_path}{Library path.}
}
\value{
Library path.
}
\description{
Creates, installs and activates
a local package library.
}
