% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_build_repo.R
\name{pkgr_build_repo}
\alias{pkgr_build_repo}
\title{Build Repository}
\usage{
pkgr_build_repo(package_files, repo_path)
}
\arguments{
\item{package_files}{Character vector of package files.}

\item{repo_path}{Repository path.}
}
\value{
Repository path.
}
\description{
Creates and populates a local CRAN-like repository.
}
