% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_build_sources.R
\name{pkgr_build_sources}
\alias{pkgr_build_sources}
\title{Build Source Packages}
\usage{
pkgr_build_sources(packages, output_path)
}
\arguments{
\item{packages}{Character vector.}

\item{output_path}{Output folder.}
}
\value{
Data frame.
}
\description{
Creates source package files for
installed packages.
}
