% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_bundle.R
\name{pkgr_bundle}
\alias{pkgr_bundle}
\title{Create Deployment Bundle}
\usage{
pkgr_bundle(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Generates deployment bundle information.
}
