% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_bundle_status.R
\name{pkgr_bundle_status}
\alias{pkgr_bundle_status}
\title{Bundle Status}
\usage{
pkgr_bundle_status(package, repo_path, bundle_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}

\item{bundle_path}{Bundle path.}
}
\value{
List.
}
\description{
Summarizes bundle readiness.
}
