% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_check.R
\name{pkgr_check}
\alias{pkgr_check}
\title{Check Package Status}
\usage{
pkgr_check(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Data frame containing package status.
}
\description{
Check whether a package is installed.
}
