% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_collect_dependencies.R
\name{pkgr_collect_dependencies}
\alias{pkgr_collect_dependencies}
\title{Collect Dependency Packages}
\usage{
pkgr_collect_dependencies(package, source_path, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{source_path}{Source package folder.}

\item{repo_path}{Repository path.}
}
\value{
Data frame.
}
\description{
Copies required dependency package files
into a repository.
}
