% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_create_library.R
\name{pkgr_create_library}
\alias{pkgr_create_library}
\title{Create Local Library}
\usage{
pkgr_create_library(library_path)
}
\arguments{
\item{library_path}{Library path.}
}
\value{
Library path.
}
\description{
Creates a local package library.
}
