% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_create_repo.R
\name{pkgr_create_repo}
\alias{pkgr_create_repo}
\title{Create Local Repository}
\usage{
pkgr_create_repo(repo_path)
}
\arguments{
\item{repo_path}{Repository path.}
}
\value{
Repository path.
}
\description{
Creates a CRAN-like repository structure.
}
\examples{
repo_path <- tempdir()
pkgr_create_repo(
  repo_path = repo_path
)

}
