% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_dependency_check.R
\name{pkgr_dependency_check}
\alias{pkgr_dependency_check}
\title{Dependency Check}
\usage{
pkgr_dependency_check(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
Data frame.
}
\description{
Checks whether dependency files
exist in a package repository.
}
