% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_dependency_gap.R
\name{pkgr_dependency_gap}
\alias{pkgr_dependency_gap}
\title{Dependency Gap}
\usage{
pkgr_dependency_gap(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
Data frame.
}
\description{
Returns missing dependency files.
}
