% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_dependency_manifest.R
\name{pkgr_dependency_manifest}
\alias{pkgr_dependency_manifest}
\title{Dependency Manifest}
\usage{
pkgr_dependency_manifest(package, output_file)
}
\arguments{
\item{package}{Package name.}

\item{output_file}{Output csv file.}
}
\value{
Data frame.
}
\description{
Creates a dependency manifest file
for package approval and deployment.
}
