% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_dependency_summary.R
\name{pkgr_dependency_summary}
\alias{pkgr_dependency_summary}
\title{Dependency Summary}
\usage{
pkgr_dependency_summary(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Summarizes dependency availability.
}
\examples{
\dontrun{
pkgr_dependency_summary(
  package = "Tplyr",
  repo_path = tempdir()
)
}

}
