% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_dependency_tree.R
\name{pkgr_dependency_tree}
\alias{pkgr_dependency_tree}
\title{Package Dependency Tree}
\usage{
pkgr_dependency_tree(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Data frame.
}
\description{
Returns recursive dependency tree.
}
