% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_deployment_report.R
\name{pkgr_deployment_report}
\alias{pkgr_deployment_report}
\title{Deployment Report}
\usage{
pkgr_deployment_report(package, repo_path, output_file)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}

\item{output_file}{Output csv file.}
}
\value{
Data frame.
}
\description{
Creates deployment readiness report.
}
