% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_deps.R
\name{pkgr_deps}
\alias{pkgr_deps}
\title{Get Package Dependencies}
\usage{
pkgr_deps(package, recursive = FALSE)
}
\arguments{
\item{package}{Package name.}

\item{recursive}{Include dependency tree.}
}
\value{
Data frame containing dependency information.
}
\description{
Returns package dependencies and installation status.
}
