% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_env.R
\name{pkgr_env}
\alias{pkgr_env}
\title{Display Environment Information}
\usage{
pkgr_env()
}
\value{
Data frame containing environment details.
}
\description{
Shows information about the current R environment.
}
