% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_export_repository.R
\name{pkgr_export_repository}
\alias{pkgr_export_repository}
\title{Export Repository Inventory}
\usage{
pkgr_export_repository(repo_path, output_file)
}
\arguments{
\item{repo_path}{Repository path.}

\item{output_file}{Output CSV file.}
}
\value{
Data frame.
}
\description{
Exports repository package inventory.
}
