% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_import_package.R
\name{pkgr_import_package}
\alias{pkgr_import_package}
\title{Import Package Files}
\usage{
pkgr_import_package(source_path, repo_path)
}
\arguments{
\item{source_path}{Source folder.}

\item{repo_path}{Repository path.}
}
\value{
Data frame.
}
\description{
Imports package files into a repository.
}
