% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_install_bundle.R
\name{pkgr_install_bundle}
\alias{pkgr_install_bundle}
\title{Install Deployment Bundle}
\usage{
pkgr_install_bundle(bundle_path, library_path, validate = TRUE)
}
\arguments{
\item{bundle_path}{Bundle path.}

\item{library_path}{Library path.}

\item{validate}{Validate bundle before installation.}
}
\value{
Installed packages.
}
\description{
Installs all package files from a bundle
into a target library.
}
