% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_install_library.R
\name{pkgr_install_library}
\alias{pkgr_install_library}
\title{Install Package to Local Library}
\usage{
pkgr_install_library(package_file, library_path)
}
\arguments{
\item{package_file}{Package file path.}

\item{library_path}{Library path.}
}
\value{
Invisible NULL.
}
\description{
Installs a package into a local library.
}
