% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_install_package.R
\name{pkgr_install_package}
\alias{pkgr_install_package}
\title{Install Package from Repository}
\usage{
pkgr_install_package(package, repo_path, library_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}

\item{library_path}{Library path.}
}
\value{
Installed packages.
}
\description{
Installs a package and its dependencies
from a RpkgR repository into a library.
}
