% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_install_repo.R
\name{pkgr_install_repo}
\alias{pkgr_install_repo}
\title{Install Package from Repository}
\usage{
pkgr_install_repo(package, repo_path, lib = NULL)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}

\item{lib}{Optional library path.}
}
\value{
Invisible NULL.
}
\description{
Installs a package from a local CRAN-like repository.
}
