% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_manifest.R
\name{pkgr_manifest}
\alias{pkgr_manifest}
\title{Create Deployment Manifest}
\usage{
pkgr_manifest(package, repo_path, output_file)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}

\item{output_file}{Output CSV file.}
}
\value{
Data frame.
}
\description{
Creates a CSV manifest for package deployment.
}
