% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_parse_deps.R
\name{pkgr_parse_deps}
\alias{pkgr_parse_deps}
\title{Parse Dependency String}
\usage{
pkgr_parse_deps(x)
}
\arguments{
\item{x}{Dependency string.}
}
\value{
Character vector.
}
\description{
Parse Dependency String
}
\keyword{internal}
