% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_parse_version_deps.R
\name{pkgr_parse_version_deps}
\alias{pkgr_parse_version_deps}
\title{Parse Dependency Versions}
\usage{
pkgr_parse_version_deps(x)
}
\arguments{
\item{x}{Dependency string.}
}
\value{
Data frame.
}
\description{
Extract package names and minimum versions.
}
\keyword{internal}
