% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_r_required.R
\name{pkgr_r_required}
\alias{pkgr_r_required}
\title{Extract Required R Version}
\usage{
pkgr_r_required(depends)
}
\arguments{
\item{depends}{Depends field from DESCRIPTION}
}
\value{
Required R version.
}
\description{
Extract Required R Version
}
\keyword{internal}
