% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_recommend.R
\name{pkgr_recommend}
\alias{pkgr_recommend}
\title{Recommend Package Version}
\usage{
pkgr_recommend(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
List containing recommendation report.
}
\description{
Generates package recommendation report.
}
