% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_remove_package.R
\name{pkgr_remove_package}
\alias{pkgr_remove_package}
\title{Remove Package from Repository}
\usage{
pkgr_remove_package(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
Removed files.
}
\description{
Removes package files from a repository.
}
