% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repo_deps.R
\name{pkgr_repo_deps}
\alias{pkgr_repo_deps}
\title{Read Package Metadata}
\usage{
pkgr_repo_deps(package_file)
}
\arguments{
\item{package_file}{Path to tar.gz package.}
}
\value{
List containing package metadata.
}
\description{
Read DESCRIPTION information from a package source archive.
}
