% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repo_status.R
\name{pkgr_repo_status}
\alias{pkgr_repo_status}
\title{Repository Status}
\usage{
pkgr_repo_status(package, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Returns repository readiness status
for a package.
}
