% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repository_cleanup.R
\name{pkgr_repository_cleanup}
\alias{pkgr_repository_cleanup}
\title{Repository Cleanup}
\usage{
pkgr_repository_cleanup(repo_path)
}
\arguments{
\item{repo_path}{Repository path.}
}
\value{
Data frame.
}
\description{
Removes older package versions
and keeps latest version only.
}
