% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repository_health.R
\name{pkgr_repository_health}
\alias{pkgr_repository_health}
\title{Repository Health}
\usage{
pkgr_repository_health(repo_path)
}
\arguments{
\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Evaluates repository completeness.
}
\examples{
\dontrun{
pkgr_repository_health(
  repo_path = tempdir()
)
}

}
