% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repository_report.R
\name{pkgr_repository_report}
\alias{pkgr_repository_report}
\title{Repository Report}
\usage{
pkgr_repository_report(repo_path, output_file)
}
\arguments{
\item{repo_path}{Repository path.}

\item{output_file}{Output csv file.}
}
\value{
Data frame.
}
\description{
Creates repository governance report.
}
