% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_repository_snapshot.R
\name{pkgr_repository_snapshot}
\alias{pkgr_repository_snapshot}
\title{Repository Snapshot}
\usage{
pkgr_repository_snapshot(repo_path, output_file)
}
\arguments{
\item{repo_path}{Repository path.}

\item{output_file}{Output csv file.}
}
\value{
Data frame.
}
\description{
Creates repository audit snapshot.
}
