% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_required_files.R
\name{pkgr_required_files}
\alias{pkgr_required_files}
\title{Required Package Files}
\usage{
pkgr_required_files(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Data frame.
}
\description{
Returns package files required for installation.
}
\examples{
pkgr_required_files(
  package = "Tplyr"
)

}
