% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_resolve_dependencies.R
\name{pkgr_resolve_dependencies}
\alias{pkgr_resolve_dependencies}
\title{Resolve Dependencies Recursively}
\usage{
pkgr_resolve_dependencies(package, recursive = TRUE)
}
\arguments{
\item{package}{Package name.}

\item{recursive}{Logical.}
}
\value{
Data frame.
}
\description{
Returns all dependencies and metadata.
}
\keyword{internal}
