% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_resolve_version.R
\name{pkgr_resolve_version}
\alias{pkgr_resolve_version}
\title{Resolve Package Version}
\usage{
pkgr_resolve_version(package, min_version, repo_path)
}
\arguments{
\item{package}{Package name.}

\item{min_version}{Minimum required version.}

\item{repo_path}{Repository path.}
}
\value{
Character string.
}
\description{
Selects the latest available version
that satisfies a minimum version.
}
