% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_scan_repo.R
\name{pkgr_scan_repo}
\alias{pkgr_scan_repo}
\title{Scan Package Repository}
\usage{
pkgr_scan_repo(repo_path)
}
\arguments{
\item{repo_path}{Path containing package files.}
}
\value{
Data frame of package files found.
}
\description{
Scan a folder for package tar.gz files.
}
