% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_sync_repository.R
\name{pkgr_sync_repository}
\alias{pkgr_sync_repository}
\title{Synchronize Repository}
\usage{
pkgr_sync_repository(package, source_repo, target_repo)
}
\arguments{
\item{package}{Package name.}

\item{source_repo}{Source repository.}

\item{target_repo}{Target repository.}
}
\value{
Data frame.
}
\description{
Copies missing dependency packages
from a source repository.
}
