% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_update_package.R
\name{pkgr_update_package}
\alias{pkgr_update_package}
\title{Update Package in Repository}
\usage{
pkgr_update_package(package_file, repo_path)
}
\arguments{
\item{package_file}{Package file.}

\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Removes older versions of a package
and adds the new version.
}
