% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_use_library.R
\name{pkgr_use_library}
\alias{pkgr_use_library}
\title{Use Local Library}
\usage{
pkgr_use_library(library_path)
}
\arguments{
\item{library_path}{Library path.}
}
\value{
Current library paths.
}
\description{
Adds a local library to .libPaths().
}
