% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_validate_bundle.R
\name{pkgr_validate_bundle}
\alias{pkgr_validate_bundle}
\title{Validate Deployment Bundle}
\usage{
pkgr_validate_bundle(package, repo_path, bundle_path)
}
\arguments{
\item{package}{Package name.}

\item{repo_path}{Package repository path.}

\item{bundle_path}{Location containing tar.gz files.}
}
\value{
Data frame.
}
\description{
Checks whether all required package files
exist in a directory.
}
