% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_validate_repository.R
\name{pkgr_validate_repository}
\alias{pkgr_validate_repository}
\title{Validate Repository}
\usage{
pkgr_validate_repository(repo_path)
}
\arguments{
\item{repo_path}{Repository path.}
}
\value{
List.
}
\description{
Performs repository validation checks.
}
\examples{
\dontrun{
pkgr_validate_repository(
  repo_path = tempdir()
)
}

}
