% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PkgR-package.R
\docType{package}
\name{RpkgR-package}
\alias{RpkgR}
\alias{RpkgR-package}
\title{RpkgR}
\description{
R Package Management for Regulated Environments
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PraneethAdidela/RpkgR}
  \item Report bugs at \url{https://github.com/PraneethAdidela/RpkgR/issues}
}

}
\author{
\strong{Maintainer}: Praneethreddy Adidela \email{praneethadidela@gmail.com}

Authors:
\itemize{
  \item Praneethreddy Adidela \email{praneethadidela@gmail.com}
}

}
\keyword{internal}
